Fixed database errors

+++ app/src/main/java/cz/vitskalicky/lepsirozvrh/database/AccountDao.kt
@@ -9,7 +9,7 @@ abstract class AccountDao {
     @Update
     abstract suspend fun updateAccount(account: Account)
     @Insert
-    abstract suspend fun insertAccount(account: Account): Int //todo test if returning primary key works as expected or change to handle rowID, see https://developer.android.com/training/data-storage/room/accessing-data#convenience-insert and https://www.sqlite.org/rowidtable.html
+    abstract suspend fun insertAccount(account: Account): Long //todo test if returning primary key works as expected or change to handle rowID, see https://developer.android.com/training/data-storage/room/accessing-data#convenience-insert and https://www.sqlite.org/rowidtable.html
     @Delete
     abstract suspend fun deleteAccount(account: Account): Int
     @Query("DELETE FROM account WHERE id = :accountId")
